// generated/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  date: any;
  money: any;
  timestamptz: any;
};
// ...
export type Insurances = {
  __typename?: 'Insurances';
  createdAt: Scalars['timestamptz'];
  description: Scalars['String'];
  endDate: Scalars['date'];
  firstInstallment: Scalars['money'];
  id: Scalars['Int'];
  price: Scalars['money'];
  reactivationDate: Scalars['date'];
  secondInstallment: Scalars['money'];
  startDate: Scalars['date'];
  suspensionDate: Scalars['date'];
  updatedAt: Scalars['timestamptz'];
  vehicleId: Scalars['Int'];
  /** An object relationship */
  vehicleOfInsurance?: Maybe<Vehicles>;
};

/** aggregated selection of "insurances" */
export type InsurancesAggregate = {
  __typename?: 'InsurancesAggregate';
  aggregate?: Maybe<InsurancesAggregateFields>;
  nodes: Array<Insurances>;
};
// ...
export type InsurancesArrRelInsertInput = {
  data: Array<InsurancesInsertInput>;
  /** upsert condition */
  onConflict?: InputMaybe<InsurancesOnConflict>;
};
// ...
export type Maintenances = {
  __typename?: 'Maintenances';
  createdAt: Scalars['timestamptz'];
  dateDone: Scalars['date'];
  description: Scalars['String'];
  id: Scalars['Int'];
  kilometers?: Maybe<Scalars['Int']>;
  name: Scalars['String'];
  updatedAt: Scalars['timestamptz'];
  vehicleId: Scalars['Int'];
  /** An object relationship */
  vehicleOfMaintenance?: Maybe<Vehicles>;
};

/** aggregated selection of "maintenances" */
export type MaintenancesAggregate = {
  __typename?: 'MaintenancesAggregate';
  aggregate?: Maybe<MaintenancesAggregateFields>;
  nodes: Array<Maintenances>;
};
// ...
export type MaintenancesArrRelInsertInput = {
  data: Array<MaintenancesInsertInput>;
  /** upsert condition */
  onConflict?: InputMaybe<MaintenancesOnConflict>;
};
// ...
export type Revisions = {
  __typename?: 'Revisions';
  createdAt: Scalars['timestamptz'];
  dateDone: Scalars['date'];
  id: Scalars['Int'];
  kilometers: Scalars['Int'];
  updatedAt: Scalars['timestamptz'];
  vehicleId?: Maybe<Scalars['Int']>;
  /** An object relationship */
  vehicleOfRevision?: Maybe<Vehicles>;
};

/** aggregated selection of "revisions" */
export type RevisionsAggregate = {
  __typename?: 'RevisionsAggregate';
  aggregate?: Maybe<RevisionsAggregateFields>;
  nodes: Array<Revisions>;
};
// ...
export type RevisionsArrRelInsertInput = {
  data: Array<RevisionsInsertInput>;
  /** upsert condition */
  onConflict?: InputMaybe<RevisionsOnConflict>;
};
// ...
export type Vehicles = {
  __typename?: 'Vehicles';
  buildDate?: Maybe<Scalars['date']>;
  createdAt: Scalars['timestamptz'];
  id: Scalars['Int'];
  image?: Maybe<Scalars['String']>;
  /** An array relationship */
  insurances: Array<Insurances>;
  /** An aggregate relationship */
  insurancesAggregate: InsurancesAggregate;
  licensePlate?: Maybe<Scalars['String']>;
  /** An array relationship */
  maintenances: Array<Maintenances>;
  /** An aggregate relationship */
  maintenancesAggregate: MaintenancesAggregate;
  name: Scalars['String'];
  /** An array relationship */
  revisions: Array<Revisions>;
  /** An aggregate relationship */
  revisionsAggregate: RevisionsAggregate;
  updatedAt: Scalars['timestamptz'];
};
// ...
export type VehiclesInsertInput = {
  buildDate?: InputMaybe<Scalars['date']>;
  createdAt?: InputMaybe<Scalars['timestamptz']>;
  id?: InputMaybe<Scalars['Int']>;
  image?: InputMaybe<Scalars['String']>;
  insurances?: InputMaybe<InsurancesArrRelInsertInput>;
  licensePlate?: InputMaybe<Scalars['String']>;
  maintenances?: InputMaybe<MaintenancesArrRelInsertInput>;
  name?: InputMaybe<Scalars['String']>;
  revisions?: InputMaybe<RevisionsArrRelInsertInput>;
  updatedAt?: InputMaybe<Scalars['timestamptz']>;
};
// ...
export type VehiclesPkColumnsInput = {
  id: Scalars['Int'];
};
// ...
export type VehiclesSetInput = {
  buildDate?: InputMaybe<Scalars['date']>;
  createdAt?: InputMaybe<Scalars['timestamptz']>;
  id?: InputMaybe<Scalars['Int']>;
  image?: InputMaybe<Scalars['String']>;
  licensePlate?: InputMaybe<Scalars['String']>;
  name?: InputMaybe<Scalars['String']>;
  updatedAt?: InputMaybe<Scalars['timestamptz']>;
};
// ...
export type CreateVehicleMutationVariables = Exact<{
  object: VehiclesInsertInput;
}>;


export type CreateVehicleMutation = { __typename?: 'mutation_root', insertVehiclesOne?: { __typename?: 'Vehicles', id: number, name: string } | null };

export type UpdateVehicleMutationVariables = Exact<{
  pkColumns: VehiclesPkColumnsInput;
  set?: InputMaybe<VehiclesSetInput>;
}>;


export type UpdateVehicleMutation = { __typename?: 'mutation_root', updateVehiclesByPk?: { __typename?: 'Vehicles', id: number } | null };

export type AllVehiclesQueryVariables = Exact<{ [key: string]: never; }>;


export type AllVehiclesQuery = { __typename?: 'query_root', vehicles: Array<{ __typename?: 'Vehicles', id: number, name: string, licensePlate?: string | null, buildDate?: any | null, image?: string | null }> };

export type SingleVehicleQueryVariables = Exact<{
  vehicleId: Scalars['Int'];
}>;


export type SingleVehicleQuery = { __typename?: 'query_root', vehiclesByPk?: { __typename?: 'Vehicles', id: number, name: string, licensePlate?: string | null, buildDate?: any | null, image?: string | null } | null };